Allow http.maxResponseSize to be Long type
maxResponseSize allows to be Long but the converter forces Integer which causes
problem when the data is set to bigger than Integer.

RB=900430
G=si-core-reviewers
R=ssheng,dhoa
A=ssheng

+++ d2/src/main/java/com/linkedin/d2/balancer/config/TransportClientPropertiesConverter.java
@@ -139,7 +139,7 @@ public class TransportClientPropertiesConverter
     if (properties.containsKey(PropertyKeys.HTTP_MAX_RESPONSE_SIZE))
     {
       config.setMaxResponseSize(coerce(properties.get(PropertyKeys.HTTP_MAX_RESPONSE_SIZE),
-          Integer.class));
+          Long.class));
     }
     if (properties.containsKey(PropertyKeys.HTTP_POOL_SIZE))
     {
